fix bug
the last bug seems to be fixed. the issue was that the last point in RS curve are wrong values, e.g. 6.7e-312, or -nan.  I didnt find out the reason; so it's temperatly fixed by check these wrong value during collision check.  A better method should be found in the future. 

Todo: 
- While the current path is valid and clear, reduce the need of finding path.  
- map-fusion needs inflation for grids in original slam output /map
- hybrid-A-star:  open and close groups should be separate for faster speed.